Better fluids working.

+++ SUN/Include/ScreenSize.h
@@ -3,6 +3,6 @@
 #include <gmtl/gmtl.h>
 using namespace gmtl;
 
-static Vec2f screenSize(800, 600);
-static Vec2f quadPos(-0.5, -0.5);
-static Vec2f quadSize(1.0, 1.0);
+static Vec2d screenSize(800, 600);
+static Vec2d quadPos(-0.5, -0.5);
+static Vec2d quadSize(1.0, 1.0);
